+++ src/smtp.rs
@@ -1,13 +1,14 @@
 use lettre;
 use std::{fmt, result};
 
-use crate::config::Account;
+use crate::config::{self, Account};
 
 // Error wrapper
 
 #[derive(Debug)]
 pub enum Error {
     TransportError(lettre::transport::smtp::Error),
+    ConfigError(config::Error),
 }
 
 impl fmt::Display for Error {
@@ -15,6 +16,7 @@ impl fmt::Display for Error {
         write!(f, "(smtp): ")?;
         match self {
             Error::TransportError(err) => err.fmt(f),
+            Error::ConfigError(err) => err.fmt(f),
         }
     }
 }
@@ -22,6 +24,12 @@ impl fmt::Display for Error {
 impl From<lettre::transport::smtp::Error> for Error {
     fn from(err: lettre::transport::smtp::Error) -> Error {
         Error::TransportError(err)
+    }
+}
+
+impl From<config::Error> for Error {
+    fn from(err: config::Error) -> Error {
+        Error::ConfigError(err)
     }
 }
 
@@ -35,7 +43,7 @@ pub fn send(account: &Account, msg: &lettre::Message) -> Result<()> {
     use lettre::Transport;
 
     lettre::transport::smtp::SmtpTransport::relay(&account.smtp_host)?
-        .credentials(account.smtp_creds())
+        .credentials(account.smtp_creds()?)
         .build()
         .send(msg)
         .map(|_| Ok(()))?