+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - List command with pagination [#19]
 - Icon in table when attachment is present [#16]
 - Multi-account [#17]
+- Password from command [#22]
 
 [unreleased]: https://github.com/soywod/himalaya
 
@@ -45,3 +46,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#17]: https://github.com/soywod/himalaya/issues/17
 [#19]: https://github.com/soywod/himalaya/issues/19
 [#21]: https://github.com/soywod/himalaya/issues/21
+[#22]: https://github.com/soywod/himalaya/issues/22